+++ lib/chef/provider/cron.rb
@@ -25,6 +25,8 @@ class Chef
     class Cron < Chef::Provider
       include Chef::Mixin::Command
 
+      provides :cron, os: "!aix"
+
       SPECIAL_TIME_VALUES = [:reboot, :yearly, :annually, :monthly, :weekly, :daily, :midnight, :hourly]
       CRON_ATTRIBUTES = [:minute, :hour, :day, :month, :weekday, :time, :command, :mailto, :path, :shell, :home, :environment]
       WEEKDAY_SYMBOLS = [:sunday, :monday, :tuesday, :wednesday, :thursday, :friday, :saturday]