+++ lib/chef/provider/group/groupmod.rb
@@ -21,6 +21,8 @@ class Chef
     class Group
       class Groupmod < Chef::Provider::Group
 
+        provides :group, os: "netbsd"
+
         def load_current_resource
           super
           [ "group", "user" ].each do |binary|