+++ lib/chef/resource/user.rb
@@ -26,6 +26,8 @@ class Chef
 
       state_attrs :uid, :gid, :home
 
+      provides :user
+
       def initialize(name, run_context=nil)
         super
         @resource_name = :user