+++ spec/functional/resource/user/dscl_spec.rb
@@ -19,9 +19,8 @@ require 'spec_helper'
 require 'chef/mixin/shell_out'
 
 metadata = {
-  :unix_only => true,
+  :mac_osx_only => true,
   :requires_root => true,
-  :provider => {:user => Chef::Provider::User::Dscl},
   :not_supported_on_mac_osx_106 => true,
 }